requirements.txt
# Vayu Dependencies
# =====================

# Speech-to-Text
faster-whisper>=1.0.0

# Text-to-Speech
kokoro-onnx>=0.4.0
onnxruntime>=1.16.0

# Audio Processing
sounddevice>=0.4.6
numpy>=1.24.0

# LLM Client (OpenRouter compatible)
openai>=1.0.0

# VAD (Voice Activity Detection)
torch>=2.0.0
torchaudio>=2.0.0

# Configuration & Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0

# Logging & Formatting
colorama>=0.4.6
